requirements.txt
flask
Flask-Session
jsonschema
pyyaml
redis
requests
